Updating Murkowsi and Duncan's very first term dates (swearing-in, end > start)
Updated Lisa Murkowski's first term start date from 2002-12-20, the date of her appointment, to 2003-01-07, the date of her swearing-in, according to Women in Congress, 1917-2007 with Biographies, By Mildred Lehmann Amer.

Updated John Duncan's first term end date from 1988-10-22 (the adjournment) to 1989-01-03 (the end of the Congress) so that the end date is not earlier than the start date which is currently set to 1988-11-08, the date he won a special election. He was probably not sworn in until the start of the next Congress since Congress appears to have been adjourned. That's a situation we don't handle well here.