add party changes
Some times Members of Congress change party during a term. We previously only had a single party field which captured their party at the end of of the term (though this wasn't always correct).

I asked @lorien to scan the roll call vote data for party changes (suggested by @joec). Then he realized the same information (and more) is in Bioguide. This adds the party change data, and also corrects several incorrect term `party` values.

The new `party_affiliations` field is present if the legislator changed party affiliation during the term. The value is a list of time periods, with `start` and `end` dates, each of which has a `party` value. The time periods cover the entire term, so the first `start` will match the term `start` and the last `end` will match the term `end` as well as the term `party`.

fixes #171