remove Thad Cochran's "first" Senate term in which the end date preceded the start date
This is a confusing one. Says bioguide

"served...until his resignation on December 26, 1978"

However due to my very old decision to end terms on the adjournment of the final session of the Congress, this term of his had already ended on Oct. 15, 1978.

"appointed by the Governor, December 27, 1978, to fill the vacancy caused by the resignation of James O. Eastland for the term ending January 3, 1979"

Congress was not actually in session during the time Cochran was filling the vacancy of the unexpired term. Since our convention in historical data is to end terms at the end of the session, we can't represent this term without putting its end date before its start date. Also since Cochran was presumably not sworn in during this time, having a term for this would violate our convention of having start dates be dates of swearing in. (If we go back and revise the historical data to match the contemporary data in ending terms at the end of the Congress, we'll need to come back and consider adding this term back. But the swearing-in problem remains.)

"elected in 1978 to the United States Senate for the term commencing January 3, 1979"

Cochran's first time actually being sworn in (presumably!) is this term.

http://bioguide.congress.gov/scripts/biodisplay.pl?index=C000567

fixes #210, albeit not very satisfyingly