remove E000142 Amos Ellmaker who has been erased from the bioguide.
Ellmaker appeared in the bioguide in 2009 but was deleted some time prior to April 2013. His biography was:

"elected as a Republican to the Fourteenth Congress, but did not qualify, having been appointed and commissioned president judge of the twelfth judicial district on July 3, 1815, and served until his resignation on December 21, 1816"

And in full:

ELLMAKER, Amos, a Representative from Pennsylvania; born at “Walnut Bottom” farm, Leacock Township, Lancaster County, Pa., February 2, 1787; attended the common schools;
was graduated from Princeton College; studied law in Lancaster, Pa., and Litchfield, Conn.; was admitted to the bar and commenced practice in Harrisburg, Pa.; deputy attorney
general for Dauphin County 1809-1815; member of the State house of representatives in 1813 and 1814; elected as a Republican to the Fourteenth Congress, but did not qualify,
having been appointed and commissioned president judge of the twelfth judicial district on July 3, 1815, and served until his resignation on December 21, 1816; attorney general
of Pennsylvania 1816-1819; moved to Lancaster, Pa., in 1821 and resumed the practice of law; again attorney general of the State in 1828 and 1829; unsuccessful candidate for Vice
President of the United States on the Anti-Masonic ticket in 1832 and for the United States Senate in 1833; continued the practice of law until his death in Lancaster, Pa.,
November 28, 1851; interment in St. James’ Episcopal Churchyard.