Wikidata IDs and updated Wikipedia article titles for all
Got Wikidata IDs for the rest of them by starting with the Wikipedia
titles, auditing them to make sure they were referring to the right
person, and then adding them in. I updated page titles for articles
that were renamed; in the future, a script will be able to do this
automatically based on the Wikidata ID. The question is to how to do
this automatically in the long term.