combine the first and middle name fields when the first name is just an initial
In these cases, the middle name is (probably) always obligatorily a part of the proper way of displaying the name when showing a first name. In other cases, the middle name is generally not used for display purposes. With this commit, the middle name field can generally be ignored for display because when it's obligatory it's now combined with the first name.