fix update_gh_pages.sh which wouldn't work if a new .yaml file was added
Because the glob was on the checked out files (not present in gh-pages) instead of on the git tree --- fixed by adding quotes --- and the glob to commit would pick up the new file before it was git-add'ed.

Also made some simplifications in the script, and use 'python3' so I can run it locally more easily.