committee_membership.py: ensure bioguide and thomas IDs are outputted in the same stable order as we have been using to prevent spurrious diffs
For some reason, running this script now suddenly reverses the thomas
and bioguide ID fields compared to the order that we have in the YAML
file. We were using a dict to pass this information, and dicts have been
unordered (though that changes in CPython 3.6, though I'm running 3.5), so
the order was only lucky to be stable until now. It's surprising we got
this far without seeing them flip.

Changing it to an OrderedDict should guarantee the order that we want.

see #546