Delete the historical record for Rep. De Nivernais who was the same person as Rep. Livernash
See https://bioguide.congress.gov/search/bio/L000365. The record for De Nivernais, which must have existed when this database was created, no longer exists as a separate entry in the Bioguide.

The defunct Bioguide ID was re-assigned to the new Rep. Donald Davis.

The deleted entry was missing a district number but the other entry has one.